Date adjustments for event based content
https://southcomm.atlassian.net/browse/CS-2187

Updated events and webinars dates to correspond with the the date range for the event and the start date for the webinar. Added <if> for one day events, plus a class to target the date for any css changes down the line.

Debra approved the setup. No date on whitepapers, start date on webcasts, and a start-end date on events (unless it is a single day event, in which case, it would only show a start date).